+++ src/main/java/roomescape/exception/exception/CustomException.java
@@ -1,5 +1,7 @@
 package roomescape.exception.exception;
 
+import roomescape.exception.dto.ErrorCode;
+
 public interface CustomException {
     ErrorCode getErrorCode();
 }
