+++ src/main/java/roomescape/exception/exception/DataReferencedException.java
@@ -1,5 +1,7 @@
 package roomescape.exception.exception;
 
+import roomescape.exception.dto.ErrorCode;
+
 public class DataReferencedException extends BaseCustomException {
     public DataReferencedException(ErrorCode errorCode) {
         super(errorCode);
