+++ src/main/java/roomescape/exception/exception/DuplicatedResourceException.java
@@ -1,5 +1,7 @@
 package roomescape.exception.exception;
 
+import roomescape.exception.dto.ErrorCode;
+
 public class DuplicatedResourceException extends BaseCustomException {
     public DuplicatedResourceException(ErrorCode errorCode) {
         super(errorCode);
