+++ src/main/java/roomescape/exception/exception/InvalidRequestException.java
@@ -1,5 +1,7 @@
 package roomescape.exception.exception;
 
+import roomescape.exception.dto.ErrorCode;
+
 public class InvalidRequestException extends BaseCustomException {
     public InvalidRequestException(ErrorCode errorCode) {
         super(errorCode);
