+++ src/main/java/roomescape/exception/exception/NotFoundResourceException.java
@@ -1,5 +1,7 @@
 package roomescape.exception.exception;
 
+import roomescape.exception.dto.ErrorCode;
+
 public class NotFoundResourceException extends BaseCustomException {
     public NotFoundResourceException(ErrorCode errorCode) {
         super(errorCode);
