+++ src/main/java/roomescape/service/ThemeService.java
@@ -9,13 +9,13 @@ import roomescape.dto.theme.AddThemeRequest;
 import roomescape.dto.theme.PopularConditionRequest;
 import roomescape.exception.exception.DataReferencedException;
 import roomescape.exception.exception.DuplicatedResourceException;
-import roomescape.exception.exception.ErrorCode;
+import roomescape.exception.dto.ErrorCode;
 import roomescape.repository.reservation.ReservationRepository;
 import roomescape.repository.theme.ThemeRepository;
 
 import java.util.List;
 
-import static roomescape.exception.exception.ErrorCode.DUPLICATED_THEME;
+import static roomescape.exception.dto.ErrorCode.DUPLICATED_THEME;
 
 @Service
 public class ThemeService {
